requirements.txt
absl-py==0.8.1
astor==0.8.1
cowsay==2.0.3
cycler==0.10.0
gast==0.3.2
grpcio==1.25.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
numpy==1.17.4
protobuf==3.11.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-Levenshtein==0.12.0
six==1.13.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.16.0
